+++ App.tsx
@@ -70,7 +70,7 @@ const App = () => {
     });
 
     const playStoreUrl = await VersionCheck.getPlayStoreUrl({
-      packageName: 'com.iconkaset.farmer',
+      packageName: 'com.iconkaset.droner',
     });
 
     const {remote} = await storeVersion({
@@ -90,6 +90,7 @@ const App = () => {
         {
           text: 'อัพเดท',
           onPress: () => {
+            mixpanel.track('กดอัพเดทแอพ')
             Linking.openURL(isIOS ? storeUrl : playStoreUrl);
             RNExitApp.exitApp();
           },