+++ src/assets/icons/icons.ts
@@ -77,5 +77,5 @@ export default {
   copyClipboard: require('./copyClipboard.png'),
   winnerCampaignBlank: require('./WinnerCampaignBlank.png'),
   checkOrange: require('./checkOrange.png'),
-  closeX : require('./closeX.png')
+  closeX: require('./closeX.png'),
 };