+++ src/assets/images/image.ts
@@ -41,6 +41,5 @@ export default {
   num1: require('./num1.png'),
   num2: require('./num2.png'),
   num3: require('./num3.png'),
-  regisSuccess : require('./regisSuccess.png')
-
+  regisSuccess: require('./regisSuccess.png'),
 };